Split transfer
